feat: replace parallelreduce test
